Add infra-accessibility-reports project
This commit adds a new `infra-accessibility-reports` project, which exists to set-up the resources in integration and staging that can be read by the `govuk-accessibility-reports-integration-data-reader` role in the Tools account.

To add some context, the `app-accessibility-reports` machine will run in the Tools account and needs access to certain S3 buckets in integration and staging. This project sets up these resources in integration and staging such that they can be read by the named role in the Tools account.